requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-pinecone==0.2.6
langchain-huggingface==0.2.0
pinecone;
sentence-transformers==4.1.0
python-dotenv==1.0.0
streamlit==1.45.1